+++ src/main/java/com/armilp/ezweight/network/WeightUpdatePacket.java
@@ -3,7 +3,11 @@ package com.armilp.ezweight.network;
 import com.armilp.ezweight.data.ItemWeightRegistry;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraftforge.network.NetworkDirection;
 import net.minecraftforge.network.NetworkEvent;
+import net.minecraftforge.server.ServerLifecycleHooks;
 
 import java.util.function.Supplier;
 
@@ -29,7 +33,15 @@ public class WeightUpdatePacket {
         context.get().enqueueWork(() -> {
             ItemWeightRegistry.setWeight(packet.itemId, packet.weight);
             ItemWeightRegistry.saveToFile(ItemWeightRegistry.getConfigFile());
+                ItemWeightRegistry.reloadFromFile();
+            MinecraftServer server = ServerLifecycleHooks.getCurrentServer();
+            if (server != null) {
+                for (ServerPlayer player : server.getPlayerList().getPlayers()) {
+                    EZWeightNetwork.CHANNEL.sendTo(new SyncALLWeightPacket(packet.itemId, packet.weight), player.connection.connection, NetworkDirection.PLAY_TO_CLIENT);
+                }
+            }
         });
         context.get().setPacketHandled(true);
     }
+
 }